Improved DRY usage
