+++ services/admin/routes/admin/dashboard/index.js
@@ -2,8 +2,10 @@ const express = require('express');
 
 const router = express.Router();
 
-const dashboard = require('./dashboard');
+const dashboard = require('./dashboard'),
+      infoEndpoint = require('./info');
 
 router.use('/', dashboard);
+router.use('/info', infoEndpoint);
 
 module.exports = router;
