+++ services/admin/routes/admin/login/index.js
@@ -2,8 +2,8 @@ const express = require('express');
 
 const router = express.Router();
 
-const login = require('./login');
-const auth = require('./auth');
+const login = require('./login'),
+      auth = require('./auth');
 
 router.use('/', login);
 router.use('/auth', auth);
